(PC-26279)[PRO] refactor: simplify shouldDisplayEACInformationSection prop
