+++ pro/src/pages/Home/venueUtils.ts
@@ -1,7 +1,11 @@
+import { isBefore, addDays } from 'date-fns'
+
 import {
+  DMSApplicationstatus,
   GetOffererResponseModel,
   GetOffererVenueResponseModel,
 } from 'apiClient/v1'
+import { getLastCollectiveDmsApplication } from 'utils/getLastCollectiveDmsApplication'
 
 export const getVirtualVenueFromOfferer = (
   offerer?: GetOffererResponseModel | null
@@ -22,3 +26,31 @@ export const hasOffererAtLeastOnePhysicalVenue = (
   offerer?: GetOffererResponseModel | null
 ): boolean =>
   offerer?.managedVenues?.some((venue) => !venue.isVirtual && venue.id) ?? false
+
+export const shouldDisplayEACInformationSectionForVenue = (
+  venue: GetOffererVenueResponseModel
+): boolean => {
+  const dmsInformations = getLastCollectiveDmsApplication(
+    venue.collectiveDmsApplications
+  )
+
+  const hasAdageIdForMoreThan30Days =
+    venue.hasAdageId &&
+    !!venue.adageInscriptionDate &&
+    isBefore(new Date(venue.adageInscriptionDate), addDays(new Date(), -30))
+
+  const hasRefusedApplicationForMoreThan30Days =
+    (dmsInformations?.state == DMSApplicationstatus.REFUSE ||
+      dmsInformations?.state == DMSApplicationstatus.SANS_SUITE) &&
+    dmsInformations.processingDate &&
+    isBefore(
+      new Date(dmsInformations?.processingDate),
+      addDays(new Date(), -30)
+    )
+
+  return (
+    Boolean(dmsInformations) &&
+    !hasAdageIdForMoreThan30Days &&
+    !hasRefusedApplicationForMoreThan30Days
+  )
+}
